requirements.txt
typer[all]==0.12.5
rich==13.9.4
pyyaml==6.0.2
GitPython==3.1.43
groq==0.11.0
python-dotenv==1.0.1
requests==2.31.0
setuptools
wheel
click<8.2.0
